[Main] Template Code 추가
